remove all dependencies in one file
