+++ notelink/api/auth/backend.py
@@ -1,6 +1,6 @@
 from fastapi_users.authentication import AuthenticationBackend
 
-from notelink.api.auth.strategy import get_database_strategy
+from notelink.api.auth.dependencies import get_database_strategy
 from notelink.core.authentication.transport import bearer_transport
