--- notelink/api/auth/strategy.py
@@ -1,23 +1,0 @@
-from typing import TYPE_CHECKING, Annotated
-
-from fastapi import Depends
-from fastapi_users.authentication.strategy.db import DatabaseStrategy
-
-from notelink.api.auth.access_token import get_access_tokens_db
-from notelink.core.settings.config import settings
-
-if TYPE_CHECKING:
-    from notelink.core.models import AccessToken
-    from fastapi_users.authentication.strategy.db import AccessTokenDatabase
-
-
-def get_database_strategy(
-    access_tokens_db: Annotated[
-        "AccessTokenDatabase[AccessToken]",
-        Depends(get_access_tokens_db),
-    ],
-) -> DatabaseStrategy:
-    return DatabaseStrategy(
-        database=access_tokens_db,
-        lifetime_seconds=settings.access_token.lifetime_seconds,
-    )
