+++ notelink/api/auth/user_routers.py
@@ -1,8 +1,8 @@
 from fastapi_users import FastAPIUsers
 
-from .backend import authentication_backend
-from .user_manager import get_user_manager
-from ...core.models import User
+from notelink.api.auth.backend import authentication_backend
+from notelink.api.auth.dependencies import get_user_manager
+from notelink.core.models import User
 from notelink.core.models.types.user_id import UserId
 
 fastapi_users = FastAPIUsers[User, UserId](
