--- notelink/api/auth/users.py
@@ -1,18 +1,0 @@
-from typing import TYPE_CHECKING, Annotated
-
-from fastapi import Depends
-
-from notelink.core.helpers import db_helper
-from notelink.core.models.user import User
-
-if TYPE_CHECKING:
-    from sqlalchemy.ext.asyncio import AsyncSession
-
-
-async def get_users_db(
-    session: Annotated[
-        "AsyncSession",
-        Depends(db_helper.session_getter),
-    ],
-):
-    yield User.get_db(session=session)
